Move top-level functions to `vectorpipe` package object
- And improve their docs. I figured out function grouping.
